requirements.txt
pytest~=7.1.2
selenium~=4.1.5
webdriver_manager~=3.7.0
Faker~=13.14.0
allure-pytest~=2.9.45